+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - CASMCMS-9557: Add basic type annotations to dbutils
     - CASMCMS-9559: Add `get_delete` method to the DB wrapper
     - CASMCMS-9561: Consolidate back-end code for multi-session-delete endpoints
+    - CASMCMS-9562: Consolidate back-end code for multi-component-patch endpoints
 
 ## [1.27.0] - 07/03/2025
 ### Changed